% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{runShinyApp}
\alias{runShinyApp}
\title{Running this function launches a Shiny web application presenting
the datasets bundled in this package, a couple of examples are included}
\usage{
runShinyApp(example)
}
\arguments{
\item{example}{the name of the bundled Shiny application example}
}
\description{
Running this function launches a Shiny web application presenting
the datasets bundled in this package, a couple of examples are included
}
