% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sebms_ggsave}
\alias{sebms_ggsave}
\title{Saves a ggplot object as a PNG file, resizing using pixel dimensions and a text scaling factor}
\usage{
sebms_ggsave(plot, width = 800, height = 500, text.factor = 1, filename)
}
\arguments{
\item{plot}{a ggplot object}

\item{width}{pixel width}

\item{height}{pixel height}

\item{text.factor}{text scaling factor (default is 1)}

\item{filename}{the path to the output file}
}
\description{
Saves a ggplot object as a PNG file, resizing using pixel dimensions and a text scaling factor
}
