% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sebms_naturum_climate}
\alias{sebms_naturum_climate}
\title{Climate data for Naturum sites from SMHI}
\usage{
sebms_naturum_climate()
}
\value{
data frame with climate data
}
\description{
Climate data for Naturum sites from SMHI
}
