% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sebms_per_update_modified}
\alias{sebms_per_update_modified}
\title{Update person data in database - setting new modified time and
database user}
\usage{
sebms_per_update_modified(my_uid)
}
\arguments{
\item{my_uid}{Internal user id in SeBMS database; use param sebms_users("nrm_msk")$uid}
}
\description{
Update person data in database - setting new modified time and
database user
}
