% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sebms_precip_temp_2015_plot}
\alias{sebms_precip_temp_2015_plot}
\title{Plot of temperature and precipitation data for 2015}
\usage{
sebms_precip_temp_2015_plot()
}
\value{
a ggplot plot object
}
\description{
Plot of temperature and precipitation data for 2015
}
