% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sebms_species_per_site_sitetype_plot}
\alias{sebms_species_per_site_sitetype_plot}
\title{Species per site and sitetype histo plot}
\usage{
sebms_species_per_site_sitetype_plot()
}
\description{
Species per site and sitetype histo plot
}
