% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sebms_species_per_year_filtered}
\alias{sebms_species_per_year_filtered}
\title{Retrieve species list per year, filter for approved ones}
\usage{
sebms_species_per_year_filtered()
}
\description{
Retrieve species list per year, filter for approved ones
}
