% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sebms_species_per_year_site_counts_filtered}
\alias{sebms_species_per_year_site_counts_filtered}
\title{List the species per year & site, add total counts 
and filter for approved species}
\usage{
sebms_species_per_year_site_counts_filtered()
}
\description{
List the species per year & site, add total counts 
and filter for approved species
}
