% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sebms_species_per_year_site_filtered}
\alias{sebms_species_per_year_site_filtered}
\title{Count the number of species per year and site 
and filter for approved ones}
\usage{
sebms_species_per_year_site_filtered()
}
\description{
Count the number of species per year and site 
and filter for approved ones
}
