% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgeomaps.R
\name{sebms_species_site_plot}
\alias{sebms_species_site_plot}
\title{Swedish map of SeBMS site/location data, making use of raster grid and point data}
\usage{
sebms_species_site_plot(occ_sp, width = 5, height = 3.5)
}
\arguments{
\item{occ_sp}{SpatialPoints with occurrence data}

\item{width}{the plot width, default 5 inches}

\item{height}{the plot height, default 3.5 inches}
}
\value{
ggplot object
}
\description{
Swedish map of SeBMS site/location data, making use of raster grid and point data
}
