% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgeomaps.R
\name{sebms_sunhours_leaflet}
\alias{sebms_sunhours_leaflet}
\title{Plot of sunhours using Leaflet}
\usage{
sebms_sunhours_leaflet(rl, ...)
}
\arguments{
\item{rl}{raster layer, if missing default is to use smhi_sunhours_rb$X2016}
}
\value{
leaflet map
}
\description{
Plot of sunhours using Leaflet
}
