% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgeomaps.R
\name{sebms_sunhours_plot}
\alias{sebms_sunhours_plot}
\title{Plot of sunhours data using ggplot}
\usage{
sebms_sunhours_plot(rl, show_legend = FALSE)
}
\arguments{
\item{rl}{raster layer, if missing default is to use smhi_sunhours_rb$X2016}

\item{show_legend}{boolean to indicate if legend should be included}
}
\value{
ggplot object
}
\description{
Plot of sunhours data using ggplot
}
