% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sebms_per_update_modified}
\alias{sebms_per_update_modified}
\title{Update person data in database - setting new modified time and
database user}
\usage{
sebms_per_update_modified(my_usr_uid = 1, target_per_uid = 0)
}
\arguments{
\item{my_usr_uid}{Internal usr_user id in SeBMS database; use param sebms_users("nrm_msk")$uid}

\item{target_per_uid}{Internal per_person id in SeBMS database}
}
\description{
Update person data in database - setting new modified time and
database user
}
